fix: inputs are now float32 instead of long

+++ neural_network_manager.py
@@ -90,7 +90,7 @@ class RepresentationNetwork(nn.Module):
     
     def forward(self, state): 
         # Ensure state tensor has a batch dimension
-        state=torch.tensor(state)
+        state=torch.tensor(state, dtype=torch.float32)
         state=state.view(-1)
         print(state)
         if state.dim() == 1:
@@ -116,8 +116,8 @@ class DynamicsNetwork(nn.Module):
     
     def forward(self, abstract_state, action):
         # Ensure state and action tensors have a batch dimension
-        abstract_state=torch.tensor(abstract_state)
-        action=torch.tensor(action)
+        abstract_state=torch.tensor(abstract_state, dtype=torch.float32)
+        action=torch.tensor(action, dtype=torch.float32)
         action=action.view(-1)
         if abstract_state.dim() == 1:
             abstract_state = abstract_state.unsqueeze(0)  # Add batch dimension
@@ -155,7 +155,7 @@ class PredictionNetwork(nn.Module):
         self.value_activation = pick_activation_func(reward_output_layer)
     
     def forward(self, abstract_state):
-        abstract_state=torch.tensor(abstract_state)
+        abstract_state=torch.tensor(abstract_state, dtype=torch.float32)
         x = self.shared_fc(abstract_state)
         
         # Policy output
